requirements.txt
appdirs==1.4.0
colorama==0.3.7
decorator==4.0.11
dj-database-url==0.4.2
Django==1.10.5
factory-boy==2.8.1
Faker==0.7.9
freezegun==0.3.8
gunicorn==19.7.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
mongoengine==0.11.0
mysqlclient==1.3.12
packaging==16.8
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.7.3.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.1.10
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.1
win-unicode-console==0.5
